ai-legal-core/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
web3
eth-account
openai
python-dotenv
